move things around. get rid of tokenizer_impls. more flexibility
